refactor: 타입명 명확히 수정
