+++ packages/client/src/hooks/useRecentlyMarket.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getRecentlyMarketList } from '@/api/coin';
+import { getRecentlyMarketList } from '@/api/market';
 
 export function useRecentlyMarket(queryString: string) {
 	const { isPending, error, data } = useQuery({
